Refactor nn functions
