+++ mellow/ops.py
@@ -1,11 +1,94 @@
 # -*- coding: utf-8 -*-
 
+import jax.numpy as np
+import jax.ops as jo
 import jax.random as random
 
-from mellow.typing import Dataset, List, Tensor
+from mellow.typing import Dataset, List, Shape, Tensor
 
 
-# -------------- data manipulators ---------------
+# ------------- network constructors -------------
+
+
+def depth(inb: int, out: int, params: Tensor) -> int:
+    """Deduces depth of hidden layer.
+
+    Args:
+        inb: Number of input nodes not excluding bias unit.
+        out: Number of output nodes.
+        params: Sequence of weights.
+
+    Returns:
+        Number of hidden nodes that, when taken with respect to `inb`
+        and `out`, represent a quasi-complete acyclic digraph with the
+        same number of arcs as weights in `params`.
+
+    Raises:
+        ValueError: If a quasi-complete acyclic digraph cannot be found
+            with the same number of arcs as weights in `params`.
+    """
+    dis = 1 + 4 * inb * (inb - 1) + 4 * out * (out - 1) + 8 * len(params)
+    hid = (1 - 2 * (inb + out) + np.sqrt(dis)) / 2
+
+    if not float(hid).is_integer():
+        msg = "{} weights insufficient for deducing depth of a [{}, :, {}] network."
+        raise ValueError(msg.format(len(params), inb, out))
+
+    return hid
+
+
+def adj_arr(shape: Shape, params: Tensor) -> Tensor:
+    """Constructs a sliced weighted acyclic-digraph adjacency matrix.
+
+    Assigns weights from `params` to an acylic-digraph adjacency matrix
+    in row-major order. Nodes are assumed to be topologically sorted.
+
+    Args:
+        shape: Tuple containing number of nodes per type.
+        params: Sequence of weights.
+
+    Returns:
+        Weighted adjacency matrix where the rows and columns represent
+        source and target nodes respectively. Only arcs with non-output
+        source and non-input target are represented.
+
+    Raises:
+        AttributeError: If weighted adjacency matrix cannot be
+            completely initialized.
+    """
+    inb, _, out = shape
+    Σ = np.sum(shape)
+
+    outbound = np.array([Σ - max(inb, n + 1) for n in range(Σ - out)])
+    arr = np.flip(np.arange(Σ - inb), 0) < outbound[:, None]
+
+    if np.count_nonzero(arr) != len(params):
+        msg = "{} weights required to initialize adj matrix of a {} network, got {}."
+        raise AttributeError(msg.format(np.count_nonzero(arr), shape, len(params)))
+
+    return jo.index_update(arr.astype(float), arr, params)
+
+
+def nd_vect(shape: Shape) -> Tensor:
+    """Constructs a node vector.
+
+    Assigns 0 to all non-output nodes except the bias unit which
+    defaults to 1. Nodes are assumed to be topologically sorted.
+
+    Args:
+        shape: Tuple containing number of nodes per type.
+
+    Returns:
+        Node vector where each element represents the stored value of a
+        non-output node.
+    """
+    Σ = np.sum(shape[0:2])
+    v = np.zeros(int(Σ), dtype=float)
+
+    return jo.index_update(v, 0, 1.0)
+
+
+# ------------- statistics operators -------------
 
 
 def shuffle(key, tensors: Dataset, axis: int = 0) -> List[Tensor]:
@@ -48,7 +131,7 @@ def batch(tensors: Dataset, step: int = 1) -> List[Tensor]:
         yield [tsr[idx:end] for tsr in tensors]
 
 
-# ------------ statistical functions -------------
+# --------------- helper functions ---------------
 
 
 def size(tensors: Dataset, axis: int = 0) -> int:
